rfctr: extract general-purpose objects to base
Extract objects from `unstructured.title` that are not particular to
that chunking strategy and place them in `unstructured.base` so they can
be re-used by other chunking strategies.

Also move the tests to `test_base.py`.
