+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.11.5-dev1
+## 0.11.5-dev2
 
 ### Enhancements
