+++ unstructured/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.11.5-dev1"  # pragma: no cover
+__version__ = "0.11.5-dev2"  # pragma: no cover
