+++ unstructured/chunking/title.py
@@ -5,26 +5,20 @@ Main entry point is the `@add_chunking_strategy()` decorator.
 
 from __future__ import annotations
 
-import collections
-import copy
-from typing import Any, DefaultDict, Dict, Iterable, Iterator, List, Optional, Tuple, cast
+from typing import Iterator, List, Optional
 
-from typing_extensions import TypeAlias
-
-from unstructured.chunking.base import ChunkingOptions
+from unstructured.chunking.base import (
+    ChunkingOptions,
+    PreChunk,
+    PreChunkBuilder,
+    PreChunkCombiner,
+    TablePreChunk,
+)
 from unstructured.documents.elements import (
-    CompositeElement,
-    ConsolidationStrategy,
     Element,
-    ElementMetadata,
-    RegexMetadata,
     Table,
-    TableChunk,
     Title,
 )
-from unstructured.utils import lazyproperty
-
-PreChunk: TypeAlias = "TablePreChunk | TextPreChunk"
 
 
 def chunk_by_title(
@@ -78,7 +72,7 @@ def chunk_by_title(
 
 def _split_elements_by_title_and_table(
     elements: List[Element], opts: ChunkingOptions
-) -> Iterator[TextPreChunk | TablePreChunk]:
+) -> Iterator[PreChunk]:
     """Implements "pre-chunker" responsibilities.
 
     A _section_ can be thought of as a "pre-chunk", generally determining the size and contents of a
@@ -102,7 +96,7 @@ def _split_elements_by_title_and_table(
 
     A Table or Checkbox element is placed into a pre-chunk by itself.
     """
-    pre_chunk_builder = TextPreChunkBuilder(opts)
+    pre_chunk_builder = PreChunkBuilder(opts)
 
     prior_element = None
 
@@ -156,396 +150,3 @@ def _metadata_differs(
     if ignore_page_numbers:
         return False
     return metadata1.page_number != metadata2.page_number
-
-
-# == PreChunks ===================================================================================
-
-
-class TablePreChunk:
-    """A pre-chunk composed of a single Table element."""
-
-    def __init__(self, table: Table, opts: ChunkingOptions) -> None:
-        self._table = table
-        self._opts = opts
-
-    def iter_chunks(self) -> Iterator[Table | TableChunk]:
-        """Split this pre-chunk into `Table` or `TableChunk` objects maxlen or smaller."""
-        text = self._table.text
-        html = self._table.metadata.text_as_html or ""
-        maxlen = self._opts.hard_max
-
-        # -- only chunk a table when it's too big to swallow whole --
-        if len(text) <= maxlen and len(html) <= maxlen:
-            yield self._table
-            return
-
-        is_continuation = False
-
-        while text or html:
-            # -- split off the next maxchars into the next TableChunk --
-            text_chunk, text = text[:maxlen], text[maxlen:]
-            table_chunk = TableChunk(text=text_chunk, metadata=copy.deepcopy(self._table.metadata))
-
-            # -- Attach maxchars of the html to the chunk. Note no attempt is made to add only the
-            # -- HTML elements that *correspond* to the TextChunk.text fragment.
-            if html:
-                html_chunk, html = html[:maxlen], html[maxlen:]
-                table_chunk.metadata.text_as_html = html_chunk
-
-            # -- mark second and later chunks as a continuation --
-            if is_continuation:
-                table_chunk.metadata.is_continuation = True
-
-            yield table_chunk
-
-            is_continuation = True
-
-
-class TextPreChunk:
-    """A sequence of elements that belong to the same semantic unit within a document.
-
-    The name "section" derives from the idea of a document-section, a heading followed by the
-    paragraphs "under" that heading. That structure is not found in all documents and actual section
-    content can vary, but that's the concept.
-
-    This object is purposely immutable.
-    """
-
-    def __init__(self, elements: Iterable[Element], opts: ChunkingOptions) -> None:
-        self._elements = list(elements)
-        self._opts = opts
-
-    def __eq__(self, other: Any) -> bool:
-        if not isinstance(other, TextPreChunk):
-            return False
-        return self._elements == other._elements
-
-    def combine(self, other_pre_chunk: TextPreChunk) -> TextPreChunk:
-        """Return new `TextPreChunk` that combines this and `other_pre_chunk`."""
-        return TextPreChunk(self._elements + other_pre_chunk._elements, opts=self._opts)
-
-    def iter_chunks(self) -> Iterator[CompositeElement]:
-        """Split this pre-chunk into one or more `CompositeElement` objects maxlen or smaller."""
-        text = self._text
-        text_len = len(text)
-        maxlen = self._opts.hard_max
-        start = 0
-        remaining = text_len
-
-        while remaining > 0:
-            end = min(start + maxlen, text_len)
-            yield CompositeElement(text=text[start:end], metadata=self._consolidated_metadata)
-            start = end
-            remaining = text_len - end
-
-    @lazyproperty
-    def text_length(self) -> int:
-        """Length of concatenated text of this pre-chunk, including separators."""
-        # -- used by pre-chunk-combiner to identify combination candidates --
-        return len(self._text)
-
-    @lazyproperty
-    def _all_metadata_values(self) -> Dict[str, List[Any]]:
-        """Collection of all populated metadata values across elements.
-
-        The resulting dict has one key for each `ElementMetadata` field that had a non-None value in
-        at least one of the elements in this pre-chunk. The value of that key is a list of all those
-        populated values, in element order, for example:
-
-            {
-                "filename": ["sample.docx", "sample.docx"],
-                "languages": [["lat"], ["lat", "eng"]]
-                ...
-            }
-
-        This preprocessing step provides the input for a specified consolidation strategy that will
-        resolve the list of values for each field to a single consolidated value.
-        """
-
-        def iter_populated_fields(metadata: ElementMetadata) -> Iterator[Tuple[str, Any]]:
-            """(field_name, value) pair for each non-None field in single `ElementMetadata`."""
-            return (
-                (field_name, value)
-                for field_name, value in metadata.known_fields.items()
-                if value is not None
-            )
-
-        field_values: DefaultDict[str, List[Any]] = collections.defaultdict(list)
-
-        # -- collect all non-None field values in a list for each field, in element-order --
-        for e in self._elements:
-            for field_name, value in iter_populated_fields(e.metadata):
-                field_values[field_name].append(value)
-
-        return dict(field_values)
-
-    @lazyproperty
-    def _consolidated_metadata(self) -> ElementMetadata:
-        """Metadata applicable to this pre-chunk as a single chunk.
-
-        Formed by applying consolidation rules to all metadata fields across the elements of this
-        pre-chunk.
-
-        For the sake of consistency, the same rules are applied (for example, for dropping values)
-        to a single-element pre-chunk too, even though metadata for such a pre-chunk is already
-        "consolidated".
-        """
-        return ElementMetadata(**self._meta_kwargs)
-
-    @lazyproperty
-    def _consolidated_regex_meta(self) -> Dict[str, List[RegexMetadata]]:
-        """Consolidate the regex-metadata in `regex_metadata_dicts` into a single dict.
-
-        This consolidated value is suitable for use in the chunk metadata. `start` and `end`
-        offsets of each regex match are also adjusted for their new positions.
-        """
-        chunk_regex_metadata: Dict[str, List[RegexMetadata]] = {}
-        separator_len = len(self._opts.text_separator)
-        running_text_len = 0
-        start_offset = 0
-
-        for element in self._elements:
-            text_len = len(element.text)
-            # -- skip empty elements like `PageBreak("")` --
-            if not text_len:
-                continue
-            # -- account for blank line between "squashed" elements, but not before first element --
-            running_text_len += separator_len if running_text_len else 0
-            start_offset = running_text_len
-            running_text_len += text_len
-
-            if not element.metadata.regex_metadata:
-                continue
-
-            # -- consolidate any `regex_metadata` matches, adjusting the match start/end offsets --
-            element_regex_metadata = copy.deepcopy(element.metadata.regex_metadata)
-            for regex_name, matches in element_regex_metadata.items():
-                for m in matches:
-                    m["start"] += start_offset
-                    m["end"] += start_offset
-                chunk_matches = chunk_regex_metadata.get(regex_name, [])
-                chunk_matches.extend(matches)
-                chunk_regex_metadata[regex_name] = chunk_matches
-
-        return chunk_regex_metadata
-
-    @lazyproperty
-    def _meta_kwargs(self) -> Dict[str, Any]:
-        """The consolidated metadata values as a dict suitable for constructing ElementMetadata.
-
-        This is where consolidation strategies are actually applied. The output is suitable for use
-        in constructing an `ElementMetadata` object like `ElementMetadata(**self._meta_kwargs)`.
-        """
-        CS = ConsolidationStrategy
-        field_consolidation_strategies = ConsolidationStrategy.field_consolidation_strategies()
-
-        def iter_kwarg_pairs() -> Iterator[Tuple[str, Any]]:
-            """Generate (field-name, value) pairs for each field in consolidated metadata."""
-            for field_name, values in self._all_metadata_values.items():
-                strategy = field_consolidation_strategies.get(field_name)
-                if strategy is CS.FIRST:
-                    yield field_name, values[0]
-                # -- concatenate lists from each element that had one, in order --
-                elif strategy is CS.LIST_CONCATENATE:
-                    yield field_name, sum(values, cast(List[Any], []))
-                # -- union lists from each element, preserving order of appearance --
-                elif strategy is CS.LIST_UNIQUE:
-                    # -- Python 3.7+ maintains dict insertion order --
-                    ordered_unique_keys = {key: None for val_list in values for key in val_list}
-                    yield field_name, list(ordered_unique_keys.keys())
-                elif strategy is CS.REGEX:
-                    yield field_name, self._consolidated_regex_meta
-                elif strategy is CS.DROP:
-                    continue
-                else:
-                    # -- not likely to hit this since we have a test in `text_elements.py` that
-                    # -- ensures every ElementMetadata fields has an assigned strategy.
-                    raise NotImplementedError(
-                        f"metadata field {repr(field_name)} has no defined consolidation strategy"
-                    )
-
-        return dict(iter_kwarg_pairs())
-
-    @lazyproperty
-    def _text(self) -> str:
-        """The concatenated text of all elements in this pre-chunk.
-
-        Each element-text is separated from the next by a blank line ("\n\n").
-        """
-        text_separator = self._opts.text_separator
-        return text_separator.join(e.text for e in self._elements if e.text)
-
-
-class TextPreChunkBuilder:
-    """An element accumulator suitable for incrementally forming a pre-chunk.
-
-    Provides monitoring properties like `.remaining_space` and `.text_length` a pre-chunker can use
-    to determine whether it should add the next element in the element stream.
-
-    `.flush()` is used to build a `TextPreChunk` object from the accumulated elements. This method
-    returns an interator that generates zero-or-one `TextPreChunk` object and is used like so:
-
-        yield from builder.flush()
-
-    If no elements have been accumulated, no `TextPreChunk` is generated. Flushing the builder
-    clears the elements it contains so it is ready to build the next text-pre-chunk.
-    """
-
-    def __init__(self, opts: ChunkingOptions) -> None:
-        self._opts = opts
-        self._separator_len = len(opts.text_separator)
-        self._elements: List[Element] = []
-
-        # -- only includes non-empty element text, e.g. PageBreak.text=="" is not included --
-        self._text_segments: List[str] = []
-        # -- combined length of text-segments, not including separators --
-        self._text_len: int = 0
-
-    def add_element(self, element: Element) -> None:
-        """Add `element` to this section."""
-        self._elements.append(element)
-        if element.text:
-            self._text_segments.append(element.text)
-            self._text_len += len(element.text)
-
-    def flush(self) -> Iterator[TextPreChunk]:
-        """Generate zero-or-one `PreChunk` object and clear the accumulator.
-
-        Suitable for use to emit a PreChunk when the maximum size has been reached or a semantic
-        boundary has been reached. Also to clear out a terminal pre-chunk at the end of an element
-        stream.
-        """
-        if not self._elements:
-            return
-        # -- clear builder before yield so we're not sensitive to the timing of how/when this
-        # -- iterator is exhausted and can add eleemnts for the next pre-chunk immediately.
-        elements = self._elements[:]
-        self._elements.clear()
-        self._text_segments.clear()
-        self._text_len = 0
-        yield TextPreChunk(elements, self._opts)
-
-    @property
-    def remaining_space(self) -> int:
-        """Maximum text-length of an element that can be added without exceeding maxlen."""
-        # -- include length of trailing separator that will go before next element text --
-        separators_len = self._separator_len * len(self._text_segments)
-        return self._opts.hard_max - self._text_len - separators_len
-
-    @property
-    def text_length(self) -> int:
-        """Length of the text in this pre-chunk.
-
-        This value represents the chunk-size that would result if this pre-chunk was flushed in its
-        current state. In particular, it does not include the length of a trailing separator (since
-        that would only appear if an additional element was added).
-
-        Not suitable for judging remaining space, use `.remaining_space` for that value.
-        """
-        # -- number of text separators present in joined text of elements. This includes only
-        # -- separators *between* text segments, not one at the end. Note there are zero separators
-        # -- for both 0 and 1 text-segments.
-        n = len(self._text_segments)
-        separator_count = n - 1 if n else 0
-        return self._text_len + (separator_count * self._separator_len)
-
-
-# == PreChunkCombiner ============================================================================
-
-
-class PreChunkCombiner:
-    """Filters pre-chunk stream to combine small pre-chunks where possible."""
-
-    def __init__(self, pre_chunks: Iterable[PreChunk], opts: ChunkingOptions):
-        self._pre_chunks = pre_chunks
-        self._opts = opts
-
-    def iter_combined_pre_chunks(self) -> Iterator[PreChunk]:
-        """Generate pre-chunk objects, combining TextPreChunk objects when they'll fit in window."""
-        accum = TextPreChunkAccumulator(self._opts)
-        combine_text_under_n_chars = self._opts.combine_text_under_n_chars
-
-        for pre_chunk in self._pre_chunks:
-            # -- start new pre-chunk under these conditions --
-            if (
-                # -- a table pre-chunk is never combined --
-                isinstance(pre_chunk, TablePreChunk)
-                # -- don't add another pre-chunk once length has reached combination soft-max --
-                or accum.text_length >= combine_text_under_n_chars
-                # -- combining would exceed hard-max --
-                or accum.remaining_space < pre_chunk.text_length
-            ):
-                yield from accum.flush()
-
-            # -- a table pre-chunk is never combined so don't accumulate --
-            if isinstance(pre_chunk, TablePreChunk):
-                yield pre_chunk
-            else:
-                accum.add_pre_chunk(pre_chunk)
-
-        yield from accum.flush()
-
-
-class TextPreChunkAccumulator:
-    """Accumulates, measures, and combines pre-chunk objects.
-
-    Provides monitoring properties `.remaining_space` and `.text_length` suitable for deciding
-    whether to add another pre-chunk.
-
-    `.flush()` is used to combine the accumulated pre-chunks into a single `TextPreChunk` object.
-    This method returns an interator that generates zero-or-one `TextPreChunk` objects and is used
-    like so:
-
-        yield from accum.flush()
-
-    If no pre-chunks have been accumulated, no `TextPreChunk` is generated. Flushing the builder
-    clears the pre-chunks it contains so it is ready to accept the next text-pre-chunk.
-    """
-
-    def __init__(self, opts: ChunkingOptions) -> None:
-        self._opts = opts
-        self._pre_chunks: List[TextPreChunk] = []
-
-    def add_pre_chunk(self, pre_chunk: TextPreChunk) -> None:
-        """Add a pre-chunk to the accumulator for possible combination with next pre-chunk."""
-        self._pre_chunks.append(pre_chunk)
-
-    def flush(self) -> Iterator[TextPreChunk]:
-        """Generate all accumulated pre-chunks as a single combined pre-chunk."""
-        pre_chunks = self._pre_chunks
-
-        # -- nothing to do if no pre-chunks have been accumulated --
-        if not pre_chunks:
-            return
-
-        # -- otherwise combine all accumulated pre-chunk into one --
-        pre_chunk = pre_chunks[0]
-        for other_pre_chunk in pre_chunks[1:]:
-            pre_chunk = pre_chunk.combine(other_pre_chunk)
-        yield pre_chunk
-
-        # -- and reset the accumulator (to empty) --
-        pre_chunks.clear()
-
-    @property
-    def remaining_space(self) -> int:
-        """Maximum size of pre-chunk that can be added without exceeding maxlen."""
-        maxlen = self._opts.hard_max
-        return (
-            maxlen
-            if not self._pre_chunks
-            # -- an additional pre-chunk will also incur an additional separator --
-            else maxlen - self.text_length - len(self._opts.text_separator)
-        )
-
-    @property
-    def text_length(self) -> int:
-        """Size of concatenated text in all pre-chunks in accumulator."""
-        n = len(self._pre_chunks)
-
-        if n == 0:
-            return 0
-
-        total_text_length = sum(s.text_length for s in self._pre_chunks)
-        total_separator_length = len(self._opts.text_separator) * (n - 1)
-        return total_text_length + total_separator_length
